AddedCoreDataManager
